separated form into its own component
