+++ app/javascript/image_sharing/components/App.js
@@ -1,5 +1,6 @@
 import React from 'react';
 import Header from './Header';
+import Form from './Form';
 import Footer from './Footer';
 
 export default function App() {
@@ -7,6 +8,7 @@ export default function App() {
     <div>
       <Header title="Tell us what you think" />
       {/* Put your components here: Flash Message, Form, Footer */}
+      <Form />
       <Footer />
     </div>
   );
